requirements.txt
Django==3.2.3
requests==2.25.1
uWSGI==2.0.19.1
PyJWT==2.1.0
